pyproject.toml
[project]
name = "phoenix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "argparse>=1.4.0",
    "asyncio>=4.0.0",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "box>=0.1.5",
    "datetime>=5.5",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "jsonpath-ng>=1.7.0",
    "logging>=0.4.9.6",
    "matplotlib>=3.10.1",
    "mlb-statsapi>=1.9.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pip>=25.1.1",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.8",
    "sentry-sdk>=2.37.0",
    "sqlalchemy>=2.0.43",
    "typing>=3.10.0.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = ["pytest>=8.3.5"]

[tool.pytest.ini_options]
pythonpath = "src"

[tool.black]
color = true
line-length = 88
target-version = ["py313"]

[tool.ruff]
extend-include = ["*.toml"]
output-format = "grouped"
show-fixes = true
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4


target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I001", "W"]
ignore = ["D100", "D101", "E501"]

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
known-first-party = ["common"]
known-local-folder = ["models"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
